Gideon : Refactored the Rule Interface and its implementing classes. MonoLingualRule is removed, its contents merged with BilingualRule. A new class RuleMethods is added to avoid code duplication in the implementation of the Rule Interface's methods. Some new Extra query methods for nonterminals etc are added, which are helpfull for computing certain discriminative features with respect to the rule labels.
